GenTeX.jl

Public methods

GenTeX.latex_im — Method

Generate an image from latex code.

GenTeX.show_example! — Method

Writes an example Markdown file containing LaTeX to a folder observed by Hugo. This is useful for debugging.

GenTeX.substitute_latex — Method

substitute_latex(md::AbstractString, scale::Number, im_dir; extra_packages="")

Substitute LaTeX in Markdown string md. The LaTeX images will be placed at im_dir/<h>.svg where h is a hash calculated over the math expression. The benefit of using a hash is that the browser downloads only one image per math expression. Image size can be tweaked by setting scale.

Private methods

GenTeX.allranges — Method

Combine the start and stop locations for regex matches with the ranges in between, that is, ensure that each position in the string is contained in a range.

GenTeX.hits — Method

Obtain a start and stop location for each regex matches in regexes.

GenTeX.match2range — Method

Convert a RegexMatch to its start and stop location in a string.

GenTeX.ranges — Method

Obtain a start and stop location for each regex match for rx.